Refactor slug management and tests
